Prepare to make HardwareKeyboardShortcutController per-user
Previously we have reused the same instance of

  InputMethodManagerService#mHardwareKeyboardShortcutController

across users, which needs to be updated before supporting concurrent
multiple users in InputMethodManagerService.

With this CL a new instance of HardwareKeyboardShortcutController will
be recreated every time the current IME user is switching.  This is an
important milestone to keep maintaining multiple instances of
HardwareKeyboardShortcutController for each user.

There must be no observable behavior change in this CL.

Bug: 309868254
Bug: 309837937
Test: presubmit
Change-Id: I679cbe81a475566d9f9651c9977023025bcae3a6