TV PiP: add ComponentName to notification
This lets the launcher get the icon drawable for the activity or app
themselves and decide on whether they want to use the monochrome version
of an adaptive icon if available since there is no way of passing a
drawable to the notification.

No changes are made to the icons that are already in the notification so
that they can continue to be used by other launchers.

Bug: 315325526
Test: manual

Change-Id: I156fc9707e845616d987beea9148af248b92e492