Collapse shortcuts with the same label and icon
At the moment apps can define shortcuts, but they have no way to
define multiple shortcuts for the same icon. If they define an
action with the same label and icon, we can assume that they are
the same and collapse them in the shortcut helper.

Flag: NONE
Test: Opened shortcut helper in Google Keep and confirmed that
some shortcuts collapsed
Bug: 318803330

Change-Id: I49544f6eb038a37a51d4380acebdc694e2929cdb