Force notifying property callback if DeviceConfig is not updated
It could be race that DeviceConfig still uses old values. However
it is not the purpose of the test to verify. So simply assume that
the property callback is called with the given values to continue
the test.

Also deflake another test that needs to wait for
performSurfacePlacement by the way.

Bug: 320310186
Bug: 319936577
Test: SplashScreenExceptionListTest#packageFromDeviceConfigIgnored
Change-Id: Ia4c5cd743b1ad3161ab2bb88ab45cc11e1ea391a