If the app bubble is in the overflow, use that instead of recreating
Previously we didn't show app bubbles in the overflow, ag/25579902
changed that but didn't update the showOrHideAppBubble code path
to account for the app bubble being in the overflow.

This CL modifies the code to check for an app bubble in the overflow
and remove it & add it back to the stack.

Flag: none
Test: atest BubblesTest BubbleDataTest
Test: manual - add an app bubble, add a chat bubble, dismiss app
               bubble, check the overflow to see the app bubble,
               then trigger the app bubble again, check the overflow
               to verify the app bubble isn't there
Bug: 314124017
Change-Id: Ie801e8e86178d471ee56f4095f60660917ca1944