Fix regression in SceneGestureHandler
This CL fixes a crash introduced in ag/25994973: if the user drags its
finger by the exact same amount as the touch slop, then the overSlop
will be 0f and we won't be able to compute in which direction the user
is dragging. This CL makes sure that the overSlop is not 0f, unless we
are intercepting a current swipe transition.

Bug: 291053278
Test: SwipeToSceneTest
Flag: N/A
Change-Id: Ia99a388a6d177fa706cc066d4a9223b8efa919dd