Automatically navigate between communal scenes.
This startable will navigate back to the blank scene when communal
becomes unvailable, such as when unlocking the device or switching to a
secondary user. It will also close communal when dream starts. Lastly,
it will ensure we default to communal when the device is docked.

Fixes: 319331389
Bug: 316415264
Test: atest CommunalSceneStartableTest
Test: docked device, verified communal is displayed by default. When
undocked, verified that it is not displayed after unlocking. Also
verified that it works with dreams.
Flag: ACONFIG com.android.systemui.communal_hub STAGING

Change-Id: I156f906bd4c38a94c1448076589841a950eea353