Make the profileApiVisibility user property a  hidden API
We added this as a system API but all the intended clients for this property at the moment can access hidden APIs.

Also removing the support_hiding_profiles flag. It was added as a read-only flag since it was used in UserTypeFactory which runs before SettingsProvider initialization. It has to be replaced by a new flag. This new flag will only be needed to guard the client side changes to exclude hidden profiles in API results.

Bug: 316362775
Change-Id: I28646309a8873301f63414252ab304dac76a1a46