uinput: delay from the end of the last delay
If we want to inject events at predictable times, then we shouldn't take
into account the current time when scheduling those injections, as that
could change depending on the timings of the process sending commands to
uinput. Instead, assume that the process sending commands intends time
offsets between injected events to be as close as possible to those
specified in the delay commands.

Bug: 310958309
Test: atest CtsInputTestCases
Change-Id: I3ceab49bd6f492244e53d3263bf10802353853c2