Don't use display orientation for reachability conditions
We should base whether to enable reachability on bounds filling parent
vertically/horizontally. Old assumption is incorrect that mismatched
orientation between parent and activity is the only way that activities
can be letterboxed. Letterboxed activities can have the same orientation
as parent but be letterboxed because of aspect ratio restrictions.

Fix: 323526309
Test: atest SizeCompatTests
Change-Id: Iadb220e5df7640d5fc6b515ed0c60b376e135919