Add featureFlag attribute for new permissions
This change will cause these feature-flagged permissions to be removed
for U QPR 2. There are many new permissions added after U. We are only
adding the featureFlag attributes 5 at a time for easier investigation
of issues that may arise. Included in this change:

- WRITE_VERIFICATION_STATE_E2EE_CONTACT_KEYS
- CAMERA_HEADLESS_SYSTEM_USER
- REGISTER_NSD_OFFLOAD_ENGINE
- QUARANTINE_APPS
- DETECT_SCREEN_RECORDING

Bug: 318014882
Test: Verified the permissions are not present in AndroidManifest.xml
in framework-res.apk.
Test: Applied change to 24Q2-release branch (U QPR 2) and flashed a
device with the changes. Verified the permissions are not present in
`adb shell pm list permissions`.

Change-Id: I0166256d3b3f2a87ee5143ed16a51663aa874715