Handle non-square pixels on External display
In case aspect ratio of the screen is different
from the aspect ratio of the resolution then the
pixels are non-square. There is usually a setting
on the display itself to force pixels to be square
but this setting is usually off by default.

Before this CL, the picture would be stretched
on the external display. With this CL, resolution
of the rendered picture is forced to be larger
by making logicaldisplay size larger. As the content
is now forced to render on a wider/higher surface then
when it is displayed on the screen it would appear
normal.

Bug: 304248677
Bug: 317363416
Test: atest LogicalDisplayTest DisplayDeviceTest
Test: try resolution aspect different from physical aspect-ratio
Change-Id: Ic437d62ffa4fa48972f6bdf3aaaba80e8b247b47