Limit the maxLines of the clipboard TextView to a minimum of 1
The TextView can crash if the maxLines is set to a negative value;
this shouldn't happen in general but is possible in edge cases/if
the view is measured at an unexpected time, due to how we're calculating
the max lines.

Bug: 317445600
Fix: 317445600
Test: tested no apparent behavior change; crash is not directly
reproducible but ensuring the maxLines is never less than 1 should
prevent the crash going forward.
Flag: NONE

Change-Id: I4da647b21d481e0baca9db991b4b6d39be56ffcb